+++ src/loaders/settings-components.tsx
@@ -5,13 +5,17 @@ interface SettingsOptionLoaderProps{
      optionType: "switch" | "input" | "selector",
      width?: number,
      noDescription?: boolean,
+     includeIcon?: boolean
 }
-export function SettingsOptionLoader({optionType, width, noDescription=false}: SettingsOptionLoaderProps){
+export function SettingsOptionLoader({optionType, width, noDescription=false, includeIcon=false}: SettingsOptionLoaderProps){
      return (
           <div className="flex flex-row items-center justify-between w-full">
-               <div className="space-y-1 w-full">
-                    <Skeleton className="h-3.5 w-2/5"/>
-                    {!noDescription && <Skeleton className="h-3.5 w-1/4"/>}
+               <div className="space-y-1 flex-1">
+                    <div className="flex items-center gap-2">
+                         {includeIcon && <Skeleton className="size-5"/>}
+                         <Skeleton className="h-4 w-2/5"/>
+                    </div>
+                    {!noDescription && <Skeleton className="h-4 w-1/4"/>}
                </div>
                {optionType==="switch" ? (
                     <Skeleton className="w-8 h-[18px]"/>
@@ -26,16 +30,24 @@ export function SettingsOptionLoader({optionType, width, noDescription=false}: S
 interface SettingsItemLoaderProps{
      className?: string,
      children: React.ReactNode,
-     includeDescription?: boolean
+     includeDescription?: boolean,
+     includeIcon?: boolean
 }
-export function SettingsItemLoader({className, children, includeDescription=false}: SettingsItemLoaderProps){
+export function SettingsItemLoader({className, children, includeDescription=false, includeIcon=false}: SettingsItemLoaderProps){
      return (
-          <div className="border border-muted flex flex-col gap-4 rounded-xl p-4">
-               <Skeleton className="h-4 w-1/3"/>
-               {includeDescription && (
-                    <Skeleton className="h-3.5 w-1/2"/>
-               )}
-               <div className={cn("space-y-3",className)}>
+          <div className="border border-muted flex flex-col gap-4 overflow-hidden rounded-xl py-4 px-3.5 bg-card">
+               <div className="rounded-t-xl">
+                    <div className="space-y-2 w-full">
+                         <div className="flex items-center gap-2">
+                              {includeIcon && <Skeleton className="size-4.5"/>}
+                              <Skeleton className="h-5 w-1/3"/>
+                         </div>
+                         {includeDescription && (
+                              <Skeleton className="h-4 w-1/2"/>
+                         )}
+                    </div>
+               </div>
+               <div className={cn("space-y-4",className)}>
                     {children}
                </div>
           </div>