+++ src/loaders/settings/file-formats.tsx
@@ -6,6 +6,14 @@ export default function FileFormatSettings(){
                <SettingsItemLoader>
                     <SettingsOptionLoader optionType="selector" width={128}/>
                </SettingsItemLoader>
+               <SettingsItemLoader>
+                    <SettingsOptionLoader optionType="switch"/>
+                    <SettingsOptionLoader optionType="switch"/>
+               </SettingsItemLoader>
+               <SettingsItemLoader>
+                    <SettingsOptionLoader optionType="switch"/>
+                    <SettingsOptionLoader optionType="switch"/>
+               </SettingsItemLoader>
           </div>
      )
 }